refactor

+++ internal/cli/model/model.go
@@ -168,14 +168,14 @@ type BlobberInfo struct {
 	Url                      string            `json:"url"`
 	Terms                    Terms             `json:"terms"`
 	Capacity                 int64             `json:"capacity"`
-	Allocated                int               `json:"allocated"`
-	LastHealthCheck          int               `json:"last_health_check"`
+	Allocated                int64             `json:"allocated"`
+	LastHealthCheck          int64             `json:"last_health_check"`
 	StakePoolSettings        StakePoolSettings `json:"stake_pool_settings"`
 	TotalStake               int64             `json:"total_stake"`
-	UsedAllocation           int               `json:"used_allocation"`
-	TotalOffers              int               `json:"total_offers"`
-	TotalServiceCharge       int               `json:"total_service_charge"`
-	UncollectedServiceCharge int               `json:"uncollected_service_charge"`
+	UsedAllocation           int64             `json:"used_allocation"`
+	TotalOffers              int64             `json:"total_offers"`
+	TotalServiceCharge       int64             `json:"total_service_charge"`
+	UncollectedServiceCharge int64             `json:"uncollected_service_charge"`
 	IsKilled                 bool              `json:"is_killed"`
 	IsShutdown               bool              `json:"is_shutdown"`
 }
